+++ lib/src/fonts.dart
@@ -3,8 +3,11 @@ class GYWFont {
   /// Index value transmitted to the device
   final int index;
 
-  /// Internal name of the font
+  /// Name of the font
   final String name;
+
+  /// Internal prefix of the font
+  final String prefix;
 
   /// Width (in pixels) of a character
   final int width;
@@ -17,6 +20,7 @@ class GYWFont {
 
   const GYWFont(
     this.name,
+    this.prefix,
     this.index, [
     this.width = 12,
     this.height = 12,
@@ -32,10 +36,10 @@ class GYWFont {
 
 /// Fonts supported by aRdent devices
 class GYWFonts {
-  static const basic = GYWFont("basic", 0, 14, 22, false);
-  static const title = GYWFont("title", 1, 20, 31, true);
-  static const bigBasic = GYWFont("big_basic", 2, 29, 47, true);
-  static const bigTitle = GYWFont("big_title", 3, 42, 71, true);
+  static const basic = GYWFont("Basic", "nba", 0, 14, 19, false);
+  static const title = GYWFont("Title", "nti", 1, 20, 28, true);
+  static const bigBasic = GYWFont("Big basic", "bba", 2, 28, 39, false);
+  static const bigTitle = GYWFont("Big title", "bti", 3, 42, 58, true);
 
   static const values = [basic, title, bigBasic, bigTitle];
 }